Import_to_neo4j/2_AEON_DT_AI_Project_Mckenney/requirements.txt
# AEON Automated Document Ingestion System
# Python Dependencies

# Core Dependencies
pyyaml>=6.0
watchdog>=3.0.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
camelot-py[cv]>=0.11.0

# DOCX Processing
mammoth>=1.6.0
python-docx>=1.1.0

# HTML Processing
beautifulsoup4>=4.12.0
html2text>=2020.1.16
markdownify>=0.11.6
lxml>=4.9.0

# NER Processing (for Week 3)
spacy>=3.7.0
# Download models with: python -m spacy download en_core_web_lg

# Neo4j Integration
neo4j>=5.14.0

# Vector Memory and Embeddings (Week 2)
qdrant-client>=1.7.0
sentence-transformers>=2.2.0

# Utilities
rich>=13.7.0
tqdm>=4.66.0
python-dateutil>=2.8.0

# Development Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
